edit user and delete added 9

+++ api/deleteuser.js
@@ -7,20 +7,27 @@ exports.handler = async (event) => {
     try {
       const { id } = JSON.parse(event.body);
 
+      if (!ObjectId.isValid(id)) {
+        return {
+          statusCode: 400,
+          body: JSON.stringify({ success: false, message: 'Invalid ID format' }),
+        };
+      }
+
       const client = await MongoClient.connect(process.env.MONGODB_URI, { useNewUrlParser: true, useUnifiedTopology: true });
       const db = client.db('customers');
       const collection = db.collection('vpn');
 
       const result = await collection.deleteOne({ _id: new ObjectId(id) });
 
+      client.close();
+
       if (result.deletedCount > 0) {
-        client.close();
         return {
           statusCode: 200,
           body: JSON.stringify({ success: true }),
         };
       } else {
-        client.close();
         return {
           statusCode: 404,
           body: JSON.stringify({ success: false, message: 'User not found' }),
